+++ libs/ledger-live-common/src/exchange/providers/swap.ts
@@ -1,5 +1,6 @@
 import { ExchangeProviderNameAndSignature } from ".";
 import { isIntegrationTestEnv } from "../swap/utils/isIntegrationTestEnv";
+import { getProvidersData } from "./getProvidersData";
 import network from "@ledgerhq/live-network";
 
 export type SwapProviderConfig = {
@@ -185,22 +186,6 @@ type CurrencyData = {
   signature: string;
 };
 
-type ProvidersDataResponse = {
-  name: string;
-  signature: string;
-  public_key: string;
-  public_key_curve: string;
-}[];
-
-type ProviderData = {
-  name: string;
-  publicKey: {
-    curve: string;
-    data: Buffer;
-  };
-  signature: Buffer;
-};
-
 let providerDataCache: Record<string, ProviderConfig & AdditionalProviderConfig> | null = null;
 
 export const getSwapProvider = async (
@@ -213,35 +198,6 @@ export const getSwapProvider = async (
   }
 
   return res[providerName.toLowerCase()];
-};
-
-function transformData(providersData: ProvidersDataResponse): Record<string, ProviderData> {
-  const transformed = {};
-  providersData.forEach(provider => {
-    const key = provider.name.toLowerCase();
-    transformed[key] = {
-      name: provider.name,
-      publicKey: {
-        curve: provider.public_key_curve,
-        data: Buffer.from(provider.public_key, "hex"),
-      },
-      signature: Buffer.from(provider.signature, "hex"),
-    };
-  });
-  return transformed;
-}
-
-export const getProvidersData = async (): Promise<Record<string, ProviderData>> => {
-  const { data: providersData } = await network<ProvidersDataResponse>({
-    method: "GET",
-    url: "https://crypto-assets-service.api.ledger.com/v1/partners",
-    params: {
-      output: "name,signature,public_key,public_key_curve",
-      service_name: "swap",
-    },
-  });
-
-  return transformData(providersData);
 };
 
 /**
@@ -285,7 +241,7 @@ export const fetchAndMergeProviderData = async () => {
 
   try {
     const [providersData, providersExtraData] = await Promise.all([
-      getProvidersData(),
+      getProvidersData("swap"),
       getProvidersCDNData(),
     ]);