+++ libs/ledgerjs/packages/hw-app-vet/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ledgerhq/hw-app-vet
 
+## 0.2.7-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`c83af75`](https://github.com/LedgerHQ/ledger-live/commit/c83af756fb388043c9f5a3862cae1231ec99a02c)]:
+  - @ledgerhq/cryptoassets@13.6.0-next.0
+
 ## 0.2.6
 
 ### Patch Changes