+++ libs/ledgerjs/packages/types-live/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ledgerhq/types-live
 
+## 6.52.0-next.0
+
+### Minor Changes
+
+- [#7740](https://github.com/LedgerHQ/ledger-live/pull/7740) [`c83af75`](https://github.com/LedgerHQ/ledger-live/commit/c83af756fb388043c9f5a3862cae1231ec99a02c) Thanks [@hzheng-ledger](https://github.com/hzheng-ledger)! - Add etherlink evm currency
+
 ## 6.51.1
 
 ### Patch Changes